third_party/blink/web_tests/platform/generic/fast/layers/opacity-change-stacking-context-expected.txt

Test that opacity changes from less than 1 to 1 don't cause layout if the layer remains a stacking context

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS internals.needsLayoutCount() is 0
PASS successfullyParsed is true

TEST COMPLETE
